chore(bidi): filter tests via preprocessSuite
