+++ tests/config/browserTest.ts
@@ -21,12 +21,10 @@ import { baseTest } from './baseTest';
 import { RunServer, RemoteServer } from './remoteServer';
 import { utils } from '../../packages/playwright-core/lib/coreBundle';
 import { isBidiChannel, parseHar } from '../config/utils';
-import { createSkipTestPredicate } from '../bidi/expectationUtil';
 import type { PageTestFixtures, PageWorkerFixtures } from '../page/pageTestApi';
 import type { RemoteServerOptions, PlaywrightServer } from './remoteServer';
 import type { BrowserContext, BrowserContextOptions, BrowserType, Page } from 'playwright-core';
 import type { Log } from '../../packages/trace/src/har';
-import type { TestInfo } from '@playwright/test';
 
 const { removeFolders, hostPlatform } = utils;
 
@@ -41,7 +39,6 @@ export type BrowserTestWorkerFixtures = PageWorkerFixtures & {
   isHeadlessShell: boolean;
   isFrozenWebkit: boolean;
   isBidi: boolean;
-  bidiTestSkipPredicate: (info: TestInfo) => boolean;
 };
 
 interface StartRemoteServer {
@@ -55,7 +52,6 @@ type BrowserTestTestFixtures = PageTestFixtures & {
   startRemoteServer: StartRemoteServer;
   contextFactory: (options?: BrowserContextOptions) => Promise<BrowserContext>;
   pageWithHar(options?: { outputPath?: string, content?: 'embed' | 'attach' | 'omit', omitContent?: boolean }): Promise<{ context: BrowserContext, page: Page, getLog: () => Promise<Log>, getZip: () => Promise<Map<string, Buffer>> }>
-  autoSkipBidiTest: void;
 };
 
 type ContextFactory = (options?: BrowserContextOptions) => Promise<{ context: BrowserContext, close: () => Promise<void> }>;
@@ -214,16 +210,6 @@ const test = baseTest.extend<BrowserTestTestFixtures & { _contextFactory: Contex
     };
     await use(pageWithHar);
   },
-
-  bidiTestSkipPredicate: [async ({ }, run) => {
-    const filter = await createSkipTestPredicate(test.info().project.name);
-    await run(filter);
-  }, { scope: 'worker' }],
-
-  autoSkipBidiTest: [async ({ bidiTestSkipPredicate }, run) => {
-    test.fixme(bidiTestSkipPredicate(test.info()), 'marked as timeout in bidi expectations');
-    await run();
-  }, { auto: true, scope: 'test' }],
 });
 
 export const playwrightTest = test;
